add sprite reading routine and change readers
Now readers automatically increment the offset depending on the
amount of bytes read instead of deferring such action to the caller.
This cleans a lot of unnecessary code on the caller part. If you need
to move the pointer backwards you can still decrement it (which is a
much less common action than incrementing it).

A whole new routine has been added to read sprites. It is fully
functional and implements the interpretation of the whole picture
format respecting its organization into POSTS.

For the time being, a list of sprites to read is manually specified.
Otherwise, there is no way to know if a LUMP is a sprite or not. Maybe
a regular expression would work here.

Also added a routine to write such sprites. That function turns the
POST-based format into a 2D texture using all the palettes and draws
them as PPM files.