add WIP of level reader
Added a level reader routine that finds all the levels on the WAD
directory given a regular expression for level LUMP naming (ExMy).

That routine also has a map of the possible LUMPs that can belong
to a map and the approriate methods to read each one of them.

The level reading method looks the directory index for the level
using the LUMP map and starts reading all the following ENTRIES
in the directory in search for the aforementioned level LUMPS. If
found, it calls the appropriate function to process them using the
function map. If not, it proceeeds to the next level.